Move search resources to their own module.
